requirements.txt
scikit-learn
matplotlib
numpy
python-mnist
